chore: Moved definition/enum/http-method.enum.mts to core/endpoint/definition/enum/http-method.enum.mts.
