+++ packages/architectura/src/core/endpoint/definition/enum/http-method.enum.mts
