+++ packages/architectura/src/definition/enum/_index.mts
@@ -1,10 +1,10 @@
+export * from '../../core/endpoint/definition/enum/http-method.enum.mjs';
 export * from '../../core/server/definition/enum/ports.enum.mjs';
 export * from '../../core/server/definition/enum/session.enum.mjs';
 export * from '../../service/logger/definition/enum/log-level.enum.mjs';
 export * from '../../service/stack-trace-parser/definition/enum/stack-trace-parser-service.enum.mjs';
 export * from '../../utility/definition/enum/time-formatting.enum.mjs';
 export * from './content-type.enum.mjs';
-export * from './http-method.enum.mjs';
 export * from './http-status-code.enum.mjs';
 export * from './millisecond.enum.mjs';
 export * from './numeric-base.enum.mjs';
